Added detection if Icarus Verilog is installed

+++ 01_combinational_logic/run_all_using_iverilog_under_linux_or_macos_brew.sh
@@ -1,5 +1,18 @@
 #!/bin/sh
 
+if ! command -v iverilog >/dev/null 2>&1
+then
+    echo "ERROR: Icarus Verilog (iverilog) is not in the path"               \
+         "or cannot be run."                                                 \
+         "See README.md file in the package directory for the instructions"  \
+         "how to install Icarus."                                            \
+          2>&1
+
+    echo "Press enter"
+    read enter
+    exit 1
+fi
+
 rm -rf log.txt
 
 for f in *.sv
@@ -10,6 +23,6 @@ do
     # gtkwave dump.vcd
 done
 
-rm -rf a.out
+rm -f a.out
 
 grep -e PASS -e FAIL -e error log.txt
